chore(metadata): move attribute validations in a separate file
Will be used by other forms in a following PR
